+++ nautilus_trader/adapters/dydx/endpoints/endpoint.py
@@ -82,7 +82,8 @@ class DYDXHttpEndpoint:
         url_path = url_path or self.url_path
         retry_name = self.name or "http_call"
 
-        async with self._retry_manager_pool as retry_manager:
+        retry_manager = await self._retry_manager_pool.acquire()
+        try:
             result: bytes | None = await retry_manager.run(
                 name=retry_name,
                 details=[url_path, str(params)],
@@ -91,5 +92,7 @@ class DYDXHttpEndpoint:
                 url_path=url_path,
                 payload=payload,
             )
+        finally:
+            await self._retry_manager_pool.release(retry_manager)
 
         return result